backend-python/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Rostislav Dugin", email = "rostislav.dugin@outlook.com"},
]
dependencies = [
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "bcrypt>=4.0.1",
    "sqlalchemy>=2.0.22",
    "uvicorn>=0.23.2",
    "alembic>=1.12.0",
    "pydantic>=2.4.2",
    "pyjwt>=2.8.0",
    "pytest>=7.4.2",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.28.0",
    "redis>=5.0.1",
    "faker>=19.11.0",
    "httpx>=0.25.0",
    "black>=23.11.0",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.1.1",
    "mypy>=1.7.1",
    "pre-commit>=3.5.0",
    "pytest-repeat>=0.9.3",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "python-telegram-bot[rate-limiter]>=20.7",
    "gunicorn>=21.2.0",
    "aiohttp>=3.9.1",
    "pytest-dotenv>=0.5.2",
    "aiofiles>=23.2.1",
    "types-aiofiles>=23.2.0.20240106",
    "python-multipart>=0.0.6",
    "psycopg>=3.2.3",
    "psycopg[binary]>=3.2.3",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.ruff]
line-length = 88
extend-select = [
  "PIE",
  "PT",
  "RET",
  "SIM",
  "ARG",
  "FIX",
  "PL",
]
select = [
  "COM812"
]

[tool.mypy]
exclude = ["venv"]

[tool.ruff.lint]
ignore=["PLR0913", "SIM105"]
[tool.pdm]
